chore: move everything in /Clusters into /Cluster
