+++ webui/react/src/pages/Cluster/ClustersOverview.tsx
@@ -8,11 +8,12 @@ import ResourcePoolDetails from 'components/ResourcePoolDetails';
 import Section from 'components/Section';
 import useFeature from 'hooks/useFeature';
 import usePermissions from 'hooks/usePermissions';
-import { ClusterOverallBar } from 'pages/Cluster/ClusterOverallBar';
-import { ClusterOverallStats } from 'pages/Cluster/ClusterOverallStats';
 import clusterStore from 'stores/cluster';
 import { ResourcePool } from 'types';
 import { useObservable } from 'utils/observable';
+
+import { ClusterOverallBar } from './ClusterOverallBar';
+import { ClusterOverallStats } from './ClusterOverallStats';
 
 const ClusterOverview: React.FC = () => {
   const resourcePools = useObservable(clusterStore.resourcePools);
