+++ webui/react/src/pages/Cluster/ClustersQueuedChart.module.scss
