+++ webui/react/src/pages/Cluster/ClustersQueuedChart.tsx
@@ -2,10 +2,10 @@ import { Radio } from 'antd';
 import React, { useMemo, useState } from 'react';
 
 import Section from 'components/Section';
-import ClusterHistoricalUsageChart from 'pages/Cluster/ClusterHistoricalUsageChart';
 import { V1RPQueueStat } from 'services/api-ts-sdk';
 import { DURATION_DAY, durationInEnglish } from 'utils/datetime';
 
+import ClusterHistoricalUsageChart from './ClusterHistoricalUsageChart';
 import css from './ClustersQueuedChart.module.scss';
 
 interface Props {
